Add missing castings... Windows type strlen() as size_t not int.
git-svn-id: file:///home/jj/hercules.svn/trunk@7497 956126f8-22a0-4046-8f4a-272fa8102e63

+++ ckddasd.c
@@ -381,7 +381,7 @@ char           *strtok_str = NULL;      /* save last position        */
             if (dev->dasdsfn)
             {
                 /* (N.B. must NOT have end quote, ONLY begin quote) */
-                int sfnlen = strlen( "sf=\"" ) + strlen( dev->dasdsfn ) + 1;
+                int sfnlen = (int)(strlen( "sf=\"" ) + strlen( dev->dasdsfn ) + 1);
                 sfname = malloc( sfnlen );
                 strlcpy( sfname, "sf=\"", sfnlen );
                 strlcat( sfname, dev->dasdsfn, sfnlen );
